(update): refresh 적용, refactoring
